Add Hitbox tag and enemy hit color feedback
Introduced a new 'Hitbox' tag in TagManager and assigned it to the relevant GameObject in the Mechanics scene. Updated BasicEnemyController to change color when hit by a 'Hitbox' collider, providing visual feedback instead of destroying the enemy. Commented out the enemy destruction in PlayerController to support this new feedback mechanism.